using stdint.h

+++ iwi.c
@@ -7,6 +7,7 @@
  * ----------------------------------------------------------------------------
  */
 
+#include <stdint.h>
 #include <stdlib.h>
 #include <stdbool.h>
 #include <math.h>
@@ -64,7 +65,7 @@ static int iwi_get_bbox_range(lua_State *L) {
   unsigned long long rangeLow  = (unsigned long long)geohash;
   unsigned long long rangeHigh = (unsigned long long)geohash;
 
-  rangeHigh |= (((unsigned long long int) 0x1 << (range)) - 1);
+  rangeHigh |= (((uint64_t) 0x1 << (range)) - 1);
   rangeLow &= BITMASK << range;
 
   lua_pushnumber(L, (double)rangeLow);
